NS&I have changed their site; attempting to update.
These are changes to make the module work with NS&I's new website, but so far it
isn't working.

I don't have time to continue trying to make this work right now.  I don't think
anyone else uses this code, to the best of my knowledge; I've certainly not had
anyone contact me to report that it no longer works.